requirements.txt
numpy
pandas
torch
torchvision
matplotlib
pyserial
pyyaml
scikit-learn
scipy
typing_extensions
bitstring
bluepy
